+++ srcs/parser_toolbox.c
@@ -6,7 +6,7 @@
 /*   By: aditsch <aditsch@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/01/25 21:10:39 by aditsch           #+#    #+#             */
-/*   Updated: 2017/02/10 15:55:17 by aditsch          ###   ########.fr       */
+/*   Updated: 2017/02/13 19:23:28 by aditsch          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,9 +36,9 @@ void		ft_init_color(t_color *c, char *str)
 	{
 		if (ft_tabstrlen(tab_str) == 3)
 		{
-			c->r = atoi(tab_str[0]);
-			c->g = atoi(tab_str[1]);
-			c->b = atoi(tab_str[2]);
+			c->r = atof(tab_str[0]);
+			c->g = atof(tab_str[1]);
+			c->b = atof(tab_str[2]);
 		}
 		ft_free_tabstr(tab_str);
 	}
